[#579] Fix gcc 5.4.0 conversion warnings in C++ runtime library

+++ compiler/extensions/cpp/runtime/src/zserio/FloatUtil.cpp
@@ -31,8 +31,8 @@ float convertUInt16ToFloat(uint16_t float16Value)
 {
     // decompose half precision float (float16)
     const uint16_t sign16Shifted = (float16Value & FLOAT16_SIGN_MASK);
-    const uint16_t exponent16 =
-            static_cast<uint16_t>(float16Value & FLOAT16_EXPONENT_MASK) >> FLOAT16_EXPONENT_BIT_POSITION;
+    const uint16_t exponent16 = static_cast<uint16_t>(
+            static_cast<uint16_t>(float16Value & FLOAT16_EXPONENT_MASK) >> FLOAT16_EXPONENT_BIT_POSITION);
     const uint16_t significand16 = (float16Value & FLOAT16_SIGNIFICAND_MASK);
 
     // calculate significand for single precision float (float32)
